+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+0.9.0
+-----
+
+* Move out apache hooks from `DjangoApp` to new `ApacheHostedDjangoApp`;
+* Rename `DjangoMultiSiteApp` into `ApacheHostedDjangoMultiSiteApp`
+
+
 0.8.5
 -----